+++ api-server/src/main/java/com/lastone/apiserver/service/recruitment/RecruitmentServiceImpl.java
@@ -1,5 +1,6 @@
 package com.lastone.apiserver.service.recruitment;
 
+import com.lastone.apiserver.constant.ImgTypes;
 import com.lastone.apiserver.exception.mypage.MemberNotFountException;
 import com.lastone.apiserver.exception.recruitment.IncorrectWriterException;
 import com.lastone.apiserver.exception.recruitment.RecruitmentStatusException;
@@ -10,14 +11,12 @@ import com.lastone.core.common.response.ErrorCode;
 import com.lastone.core.domain.gym.Gym;
 import com.lastone.core.domain.member.Member;
 import com.lastone.core.domain.recruitment.Recruitment;
-import com.lastone.core.domain.recruitment.RecruitmentStatus;
 import com.lastone.core.domain.recruitment_img.RecruitmentImg;
 import com.lastone.core.dto.gym.GymDto;
 import com.lastone.core.dto.recruitment.RecruitmentDetailDto;
 import com.lastone.core.dto.recruitment.RecruitmentListDto;
 import com.lastone.core.dto.recruitment.RecruitmentRequestDto;
 import com.lastone.core.dto.recruitment.RecruitmentSearchCondition;
-import com.lastone.core.dto.recruitment.StartedAtDto;
 import com.lastone.core.mapper.mapper.GymMapper;
 import com.lastone.core.repository.gym.GymRepository;
 import com.lastone.core.repository.member.MemberRepository;
@@ -32,26 +31,31 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.ObjectUtils;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
+import java.util.Objects;
+import java.util.stream.Collectors;
 
 
 @Slf4j
 @Service
-@Transactional
+@Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class RecruitmentServiceImpl implements RecruitmentService {
 
     private static final int IMG_MAX_COUNT = 3;
+
     private final S3Service s3Service;
+
     private final RecruitmentRepository recruitmentRepository;
+
     private final RecruitmentImgRepository recruitmentImgRepository;
+
     private final GymRepository gymRepository;
+
     private final MemberRepository memberRepository;
+
     private final SbdRepository sbdRepository;
+
     private final GymMapper gymMapper;
 
     @Override
@@ -61,10 +65,7 @@ public class RecruitmentServiceImpl implements RecruitmentService {
 
     @Override
     public RecruitmentDetailDto getDetail(Long recruitmentId) {
-        RecruitmentDetailDto recruitmentDetailDto = recruitmentRepository.getDetailDto(recruitmentId);
-        if (recruitmentDetailDto == null) {
-            throw new RecruitmentNotFoundException();
-        }
+        RecruitmentDetailDto recruitmentDetailDto = recruitmentRepository.getDetailDto(recruitmentId).orElseThrow(RecruitmentNotFoundException::new);
         recruitmentDetailDto.setSbdDto(sbdRepository.findLatestRecordByMemberId(recruitmentDetailDto.getMemberId()));
         return recruitmentDetailDto;
     }
@@ -74,44 +75,31 @@ public class RecruitmentServiceImpl implements RecruitmentService {
         return recruitmentRepository.getListDtoInMainPage();
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
-    public void createRecruitment(Long memberId, RecruitmentRequestDto recruitmentCreateDto, List<MultipartFile> imgFiles) throws IOException {
-
-        Gym gym = findGym(recruitmentCreateDto.getGym());
-
+    public void createRecruitment(Long memberId, RecruitmentRequestDto recruitmentRequestDto, List<MultipartFile> imgFiles) throws IOException {
+        Gym gym = findGym(recruitmentRequestDto.getGym());
         Member member = memberRepository.findById(memberId).orElseThrow(MemberNotFountException::new);
+        Recruitment recruitment = Recruitment.create(member, gym, recruitmentRequestDto);
 
-        Recruitment recruitment = Recruitment.builder()
-                .member(member)
-                .gym(gym)
-                .title(recruitmentCreateDto.getTitle())
-                .workoutPart(recruitmentCreateDto.getWorkoutPart())
-                .description(recruitmentCreateDto.getDescription())
-                .startedAt(startedAtToLocalDateTime(recruitmentCreateDto.getStartedAt()))
-                .recruitmentStatus(RecruitmentStatus.RECRUITING)
-                .preferGender(recruitmentCreateDto.getPreferGender())
-                .build();
-
-        // List에 파일이 있지만 실제로 비어있는 경우 때문에 로직 수정
-        if (imgFileIsNotEmpty(imgFiles)) {
+        if (imgFileIsExist(imgFiles)) {
             recruitment.setImgFiles(saveRecruitmentImg(imgFiles));
         }
         recruitmentRepository.save(recruitment);
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
-    public void updateRecruitment(Long recruitmentId, Long memberId, RecruitmentRequestDto recruitmentUpdateDto, List<MultipartFile> imgFiles) throws IOException {
+    public void updateRecruitment(Long recruitmentId, Long memberId, RecruitmentRequestDto recruitmentRequestDto, List<MultipartFile> imgFiles) throws IOException {
         Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(RecruitmentNotFoundException::new);
         validateWriterAndMember(recruitment.getMember().getId(), memberId);
-        recruitment.update(recruitmentUpdateDto);
-        updateGym(recruitment, recruitmentUpdateDto.getGym());
-        updateStartedAt(recruitment, recruitmentUpdateDto.getStartedAt());
-        if (imgFileIsNotEmpty(imgFiles)) {
-            updateImgFiles(recruitment, imgFiles);
-        }
-
+        /* 모집글 수정 작업 */
+        recruitment.update(recruitmentRequestDto);
+        updateGym(recruitment, recruitmentRequestDto.getGym());
+        updateImgFiles(recruitment, imgFiles);
     }
 
+    @Transactional(rollbackFor = Exception.class)
     @Override
     public void deleteRecruitment(Long recruitmentId, Long memberId) {
         Recruitment recruitment = recruitmentRepository.findById(recruitmentId).orElseThrow(RecruitmentNotFoundException::new);
@@ -124,70 +112,51 @@ public class RecruitmentServiceImpl implements RecruitmentService {
     }
 
     private void updateImgFiles(Recruitment recruitment, List<MultipartFile> imgFiles) throws IOException {
-        deleteImgFile(recruitment.getRecruitmentImgs());
-        List<RecruitmentImg> recruitmentImgs = saveRecruitmentImg(imgFiles);
-        recruitment.updateImg(recruitmentImgs);
-    }
-
-    private void deleteImgFile(List<RecruitmentImg> recruitmentImgs) {
-        for (RecruitmentImg recruitmentImg : recruitmentImgs) {
-            s3Service.delete(recruitmentImg.getImgUrl());
+        if (imgFileIsExist(imgFiles)) {
+            return;
         }
+        deleteImgFile(recruitment.getRecruitmentImgs());
+        List<RecruitmentImg> recruitmentImgList = saveRecruitmentImg(imgFiles);
+        recruitment.updateImg(recruitmentImgList);
     }
 
-    private boolean imgFileIsNotEmpty(List<MultipartFile> imgFiles) {
+    private void deleteImgFile(List<RecruitmentImg> recruitmentImgList) {
+        recruitmentImgList.forEach(recruitmentImg -> s3Service.delete(recruitmentImg.getImgUrl()));
+    }
+
+    private boolean imgFileIsExist(List<MultipartFile> imgFiles) {
         if (ObjectUtils.isEmpty(imgFiles)) {
             return false;
         }
-        for (MultipartFile imgFile : imgFiles) {
-            if (!ObjectUtils.isEmpty(imgFile)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private void updateStartedAt(Recruitment recruitment, StartedAtDto startedAt) {
-        recruitment.updateStartedAt(startedAtToLocalDateTime(startedAt));
+        /* 각각의 이미지 파일 형식 검증 */
+        imgFiles.forEach(imgFile -> ImgTypes.isSupport(Objects.requireNonNull(imgFile.getOriginalFilename())));
+        return true;
     }
 
     private void updateGym(Recruitment recruitment, GymDto gymDto) {
         Gym gym = findGym(gymDto);
-        if (recruitment.getGym().equals(gym)) {
-            return;
+        if (!recruitment.getGym().equals(gym)) {
+            recruitment.updateGym(gym);
         }
-        recruitment.updateGym(gym);
     }
 
     private void validateWriterAndMember(Long writerId, Long memberId) {
-        if (writerId.equals(memberId)) {
-            return;
+        if (!writerId.equals(memberId)) {
+            throw new IncorrectWriterException();
         }
-        throw new IncorrectWriterException();
-
     }
 
     private List<RecruitmentImg> saveRecruitmentImg(List<MultipartFile> imgFiles) throws IOException {
         validateImgCount(imgFiles);
         List<String> imgUrls = s3Service.upload(imgFiles);
-        List<RecruitmentImg> recruitmentImgs = new ArrayList<>();
-        for (String imgUrl : imgUrls) {
-            recruitmentImgs.add(recruitmentImgRepository.save(new RecruitmentImg(imgUrl)));
-        }
-        return recruitmentImgs;
-    }
-
-    private LocalDateTime startedAtToLocalDateTime(StartedAtDto startedAt) {
-        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy.MM.dd a h:mm", Locale.KOREAN);
-        return LocalDateTime.parse(startedAt.toTime(), formatter);
+        return imgUrls.stream()
+                .map(imgUrl -> recruitmentImgRepository.save(new RecruitmentImg(imgUrl)))
+                .collect(Collectors.toList());
     }
 
     private Gym findGym(GymDto gymdto) {
         Gym gym = gymRepository.findByGymDto(gymdto);
-        if (gym != null) {
-            return gym;
-        }
-        return gymRepository.save(gymMapper.toEntity(gymdto)); // 없으면 헬스장 새로 생성 후 반환
+        return Objects.requireNonNullElseGet(gym, () -> gymRepository.save(gymMapper.toEntity(gymdto)));
     }
 
     private void validateImgCount(List<MultipartFile> imgFiles) {
